+++ leia-models/src/main/java/com/grookage/leia/models/request/SearchRequest.java
@@ -16,6 +16,7 @@
 
 package com.grookage.leia.models.request;
 
+import com.fasterxml.jackson.annotation.JsonAlias;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.grookage.leia.models.schema.engine.SchemaState;
 import lombok.AllArgsConstructor;
@@ -32,7 +33,8 @@ import java.util.Set;
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class SearchRequest {
     @Builder.Default
-    private Set<String> orgs = Set.of();
+    @JsonAlias({"orgs"})
+    private Set<String> orgIds = Set.of();
     @Builder.Default
     private Set<String> tenants = Set.of();
     @Builder.Default
@@ -41,4 +43,5 @@ public class SearchRequest {
     private Set<String> schemaNames = Set.of();
     @Builder.Default
     private Set<SchemaState> states = Set.of();
+
 }